css rendering improvements; excel helper seperated into a file; visual improvements
eliminated the massive data-uri embeds and dramatically decreased the
CSS parse time.

Switched to cards in Bootstrap 4.

Some excel generic files were seperated into their own file.
