+++ __mocks__/jquery.js
@@ -1,0 +1,1 @@
+global.jQuery = global.$ = require('jquery')
