+++ export.html
@@ -1,31 +1,34 @@
 <!DOCTYPE html>
 <html lang="en"><head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-  <meta name="description" content="Easily export all your report suites config to Excel.">
-  <meta name="author" content="Gene Jones">
-  <meta name="keywords" content="Adobe Analytics, Reportsuites.com, Excel export, excel output, grading">
-  <!-- All the favicon info -->
-  <link rel="apple-touch-icon" sizes="180x180" href="/assets/icons/apple-touch-icon.png">
-  <link rel="icon" type="image/png" sizes="32x32" href="/assets/icons/favicon-32x32.png">
-  <link rel="icon" type="image/png" sizes="16x16" href="/assets/icons/favicon-16x16.png">
-  <link rel="manifest" href="/assets/icons/manifest.json">
-  <link rel="mask-icon" href="/assets/icons/icon.svg" color="#da532c">
-  <link rel="shortcut icon" href="/assets/icons/favicon.ico">
-  <meta name="msapplication-config" content="/assets/icons/browserconfig.xml">
-  <meta name="theme-color" content="#93b5b5">
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+	<title>Configurations Export to Excel</title>
+	<meta name="description" content="Easily export all your report suites config to Excel.">
+	<meta name="author" content="Gene Jones">
+	<meta name="keywords" content="Adobe Analytics, Reportsuites.com, Excel export, excel output, grading">
+	<!-- All the favicon info -->
+	<link rel="apple-touch-icon" sizes="180x180" href="/assets/icons/apple-touch-icon.png">
+	<link rel="icon" type="image/png" sizes="32x32" href="/assets/icons/favicon-32x32.png">
+	<link rel="icon" type="image/png" sizes="16x16" href="/assets/icons/favicon-16x16.png">
+	<link rel="manifest" href="/assets/icons/manifest.json">
+	<link rel="mask-icon" href="/assets/icons/icon.svg" color="#da532c">
+	<link rel="shortcut icon" href="/assets/icons/favicon.ico">
+	<meta name="msapplication-config" content="/assets/icons/browserconfig.xml">
+	<meta name="theme-color" content="#93b5b5">
 	<!-- Open Graph data -->
-  <meta property="og:title" content="Reportsuites.com: Tools for Adobe Analytics" />
-  <meta property="og:type" content="article" />
-  <meta property="og:url" content="https://reportsuites.com/" />
-  <meta property="og:image" content="https://reportsuites.com/assets/icons/mstile-150x150.png" />
-  <meta property="og:description" content="Tackling the details of your reportsuites." />
-  <meta property="og:site_name" content="Reportsuites.com" />
-  <title>Configurations Export to Excel</title>
+	<meta property="og:title" content="Reportsuites.com: Tools for Adobe Analytics" />
+	<meta property="og:type" content="article" />
+	<meta property="og:url" content="https://reportsuites.com/" />
+	<meta property="og:image" content="https://reportsuites.com/assets/icons/mstile-150x150.png" />
+	<meta property="og:description" content="Tackling the details of your reportsuites." />
+	<meta property="og:site_name" content="Reportsuites.com" />
 
+	<link rel="dns-prefetch" href="//fonts.googleapis.com/">
+	<link rel="dns-prefetch" href="//www.googletagmanager.com">
+	<link rel="dns-prefetch" href="//maxcdn.bootstrapcdn.com">
     <!-- Bootstrap core CSS -->
     <!-- Latest compiled and minified CSS -->
-    <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.3/css/bootstrap.min.css" integrity="sha384-Zug+QiDoJOrZ5t4lssLdxGhVrurbmBWopoEl+M6BdEfwnCJZtKxi1KgxUyJq13dy" crossorigin="anonymous">
 	<link href="https://fonts.googleapis.com/css?family=Allerta" rel="stylesheet">
 	<link href="https://fonts.googleapis.com/css?family=Crimson+Text" rel="stylesheet">
     <!-- Custom styles for this template -->
@@ -40,8 +43,7 @@
 	'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
 	})(window,document,'script','dataLayer','GTM-NG5QWXV');</script>
 	<!-- End Google Tag Manager -->
-	
-	 <!-- JavaScript all deferred, will work on IE10+ and all major current browsers -->
+	<!-- JavaScript all deferred, will work on IE10+ and all major current browsers -->
     <script defer src="assets/js/essentials.js"></script>
 	<script defer src="assets/js/jszip.min.js"></script>
 	<script defer src="assets/js/xmlbuilder.min.js"></script>
