+++ grader.html
@@ -22,7 +22,7 @@
   <meta property="og:description" content="Tackling the details of your reportsuites." />
   <meta property="og:site_name" content="Reportsuites.com" />
 
-  <title>Gradin</title>
+  <title>Grading your tags</title>
 
   <!-- Bootstrap core CSS -->
   <!-- Latest compiled and minified CSS -->
