+++ sitemap.xml
@@ -3,19 +3,19 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
     <loc>https://reportsuites.com/</loc>
-    <lastmod>2018-01-06</lastmod>
+    <lastmod>2018-01-15</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.7</priority>
   </url>
   <url>
     <loc>https://reportsuites.com/export.html</loc>
-    <lastmod>2018-01-06</lastmod>
+    <lastmod>2018-01-15</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
   </url>
     <url>
     <loc>https://reportsuites.com/grader.html</loc>
-    <lastmod>2018-01-06</lastmod>
+    <lastmod>2018-01-15</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.5</priority>
   </url>
